feat(common): introduce generic `ProviderBuilder::from_config` method (#13268)
- keep the existing helpers that erase generic to avoid breaking change
