+++ crates/cli/src/utils/mod.rs
@@ -115,33 +115,8 @@ pub fn get_provider_with_curl(config: &Config, curl_mode: bool) -> Result<RetryP
 /// Returns a [ProviderBuilder] instantiated using [Config] values.
 ///
 /// Defaults to `http://localhost:8545` and `Mainnet`.
-///
-/// When `curl_mode` is true, the provider will print equivalent curl commands
-/// to stdout instead of executing RPC requests.
 pub fn get_provider_builder(config: &Config, curl_mode: bool) -> Result<ProviderBuilder> {
-    let url = config.get_rpc_url_or_localhost_http()?;
-    let mut builder = ProviderBuilder::new(url.as_ref());
-
-    builder = builder.accept_invalid_certs(config.eth_rpc_accept_invalid_certs);
-    builder = builder.curl_mode(curl_mode);
-
-    if let Ok(chain) = config.chain.unwrap_or_default().try_into() {
-        builder = builder.chain(chain);
-    }
-
-    if let Some(jwt) = config.get_rpc_jwt_secret()? {
-        builder = builder.jwt(jwt.as_ref());
-    }
-
-    if let Some(rpc_timeout) = config.eth_rpc_timeout {
-        builder = builder.timeout(Duration::from_secs(rpc_timeout));
-    }
-
-    if let Some(rpc_headers) = config.eth_rpc_headers.clone() {
-        builder = builder.headers(rpc_headers);
-    }
-
-    Ok(builder)
+    ProviderBuilder::from_config(config).map(|builder| builder.curl_mode(curl_mode))
 }
 
 pub async fn get_chain<P>(chain: Option<Chain>, provider: P) -> Result<Chain>
